output/gateway/jwt-issuer/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# JWT handling
PyJWT==2.8.0

# HTTP client for health checks
httpx==0.26.0

# Data validation
pydantic==2.5.3

# Logging
python-json-logger==2.0.7
